python_code/redeemx_be_app-uat/requirements.txt
fastapi[standard]==0.115.6
sqlmodel==0.0.22
pymysql==1.1.1
passlib==1.7.4
bcrypt==4.0.1
cryptography==44.0.0
PyJWT==2.10.1
fastapi-pagination==0.12.34
fastapi-mail==1.4.2
alembic==1.11.3
websockets==11.0.3
python-dotenv==1.0.0
qrcode==7.3.1
pillow==11.1.0
loguru==0.7.0
itsdangerous==2.2.0
pandas==2.2.3
scipy==1.15.1
xlrd==2.0.1
